backend/requirements.txt
django==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
python-dotenv==1.1.0
